fix duplicated messages on bli api ai translator. Create Message Descriptor concept for indexed messages
