+++ example/Api.Ai.Csharp.Frameworks.Example.Bot.Application/Web.config
@@ -13,7 +13,7 @@
     <add key="exceptionAiMessage" value="Oops! Estou com alguns problemas por aqui. Será que podemos nos falar mais tarde? Obrigada!" />
     <add key="defaultAiMessage" value="Você deve ter enviado uma mensagem em um formato diferente do que espero ou então, eu ainda não sei sobre isso :(" />
 
-    <add key="apiAiKey" value="01c9e19e00ef4bef82d5503cc48008bd" />
+    <add key="apiAiKey" value="543b9445d21a4d1bb79c3f569f8d4827" />
   </appSettings>
   <!--
     For a description of web.config changes see http://go.microsoft.com/fwlink/?LinkId=235367.
