+++ example/Api.Ai.Csharp.Frameworks.Example.sln
@@ -1,11 +1,23 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Example.Bot.Application", "Api.Ai.Csharp.Frameworks.Example.Bot.Application\Api.Ai.Csharp.Frameworks.Example.Bot.Application.csproj", "{76AC0CE2-DB7A-4B89-9A0F-4A6B592DC345}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Example.Blip.Ai", "Api.Ai.Csharp.Frameworks.Example.Blip.Ai\Api.Ai.Csharp.Frameworks.Example.Blip.Ai.csproj", "{6E8515DF-3A4B-4C22-8E4F-B261ED59E88F}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ReferenceProjects", "ReferenceProjects", "{8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Domain.DataTransferObject", "..\src\Domain\Api.Ai.Csharp.Frameworks.Domain.DataTransferObject\Api.Ai.Csharp.Frameworks.Domain.DataTransferObject.csproj", "{5D32573C-3077-4ECE-9EA2-B784FC36B398}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Domain.Enum", "..\src\Domain\Api.Ai.Csharp.Frameworks.Domain.Enum\Api.Ai.Csharp.Frameworks.Domain.Enum.csproj", "{8EDE3190-BCDB-43A7-9EA7-E32561C90C91}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Domain.Service", "..\src\Domain\Api.Ai.Csharp.Frameworks.Domain.Service\Api.Ai.Csharp.Frameworks.Domain.Service.csproj", "{61D647D7-8EB7-48B8-8C13-E983C41D3CD4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.Blip.Ai", "..\src\Infrastructure\Api.Ai.Csharp.Frameworks.Blip.Ai\Api.Ai.Csharp.Frameworks.Blip.Ai.csproj", "{82DB0D19-744D-44FB-903F-F4AABC3BF833}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Ai.Csharp.Frameworks.BotFramework", "..\src\Infrastructure\Api.Ai.Csharp.Frameworks.BotFramework\Api.Ai.Csharp.Frameworks.BotFramework.csproj", "{84C74B76-3B1E-4D32-9327-82C9587FEF32}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,8 +33,38 @@ Global
 		{6E8515DF-3A4B-4C22-8E4F-B261ED59E88F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E8515DF-3A4B-4C22-8E4F-B261ED59E88F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6E8515DF-3A4B-4C22-8E4F-B261ED59E88F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5D32573C-3077-4ECE-9EA2-B784FC36B398}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5D32573C-3077-4ECE-9EA2-B784FC36B398}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5D32573C-3077-4ECE-9EA2-B784FC36B398}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5D32573C-3077-4ECE-9EA2-B784FC36B398}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8EDE3190-BCDB-43A7-9EA7-E32561C90C91}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8EDE3190-BCDB-43A7-9EA7-E32561C90C91}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8EDE3190-BCDB-43A7-9EA7-E32561C90C91}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8EDE3190-BCDB-43A7-9EA7-E32561C90C91}.Release|Any CPU.Build.0 = Release|Any CPU
+		{61D647D7-8EB7-48B8-8C13-E983C41D3CD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{61D647D7-8EB7-48B8-8C13-E983C41D3CD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{61D647D7-8EB7-48B8-8C13-E983C41D3CD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{61D647D7-8EB7-48B8-8C13-E983C41D3CD4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{82DB0D19-744D-44FB-903F-F4AABC3BF833}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{82DB0D19-744D-44FB-903F-F4AABC3BF833}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{82DB0D19-744D-44FB-903F-F4AABC3BF833}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{82DB0D19-744D-44FB-903F-F4AABC3BF833}.Release|Any CPU.Build.0 = Release|Any CPU
+		{84C74B76-3B1E-4D32-9327-82C9587FEF32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{84C74B76-3B1E-4D32-9327-82C9587FEF32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{84C74B76-3B1E-4D32-9327-82C9587FEF32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{84C74B76-3B1E-4D32-9327-82C9587FEF32}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{5D32573C-3077-4ECE-9EA2-B784FC36B398} = {8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}
+		{8EDE3190-BCDB-43A7-9EA7-E32561C90C91} = {8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}
+		{61D647D7-8EB7-48B8-8C13-E983C41D3CD4} = {8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}
+		{82DB0D19-744D-44FB-903F-F4AABC3BF833} = {8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}
+		{84C74B76-3B1E-4D32-9327-82C9587FEF32} = {8F5B8BE4-9278-418A-9A7A-7A9F0D14E50F}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D063C944-9AC0-4083-8458-852595717C08}
 	EndGlobalSection
 EndGlobal
