+++ src/Domain/Api.Ai.Csharp.Frameworks.Domain.DataTransferObject/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Api.Ai.Csharp" version="1.0.2.4" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.3" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net461" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net461" />
+</packages>
