+++ src/Infrastructure/Api.Ai.Csharp.Frameworks.Blip.Ai/Api.Ai.Csharp.Blip.Ai.nuspec
@@ -2,7 +2,7 @@
 <package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
     <metadata>
         <id>Api.Ai.Csharp.Blip.Ai</id>
-        <version>1.0.0.8</version>
+        <version>1.0.1.0</version>
         <title>Api.Ai.Csharp.Blip.Ai</title>
         <authors>Bruno Brandes</authors>
         <owners>Bruno Brandes</owners>
